requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
future==0.18.2
psycopg2==2.8.5
pytz==2020.1
Serializer==0.2.1
serializers==0.2.4
simplejson==3.17.2
sqlparse==0.3.1
